Move DIFPtr
